requirements.txt
requests==2.32.3
websockets==12.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1
holidays==0.67
